Add error handling to RT decode and generation
The call to JWT::decode on the provided refresh token is now surrounded
by a try/catch block that sends the user back to a generic 'log in again'
message to provide as few details as possible to bad actors.
The if statement with conditions for generating a new JWT has been
expanded to check if the token and user exist in the db, and if the
token is associated with the provided user. Otherwise, it sends the user
back to a generic 'log in again' message to provide as few details as
possible to bad actors.